requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
psycopg2-binary>=2.9
sqlalchemy>=2.0
dbt-postgres>=1.7
streamlit>=1.33
plotly>=5.0
matplotlib>=3.8
seaborn>=0.13
faker>=20.0
joblib>=1.3
python-dotenv>=1.0
pytest>=7.0